+++ read_header_command.go
@@ -57,8 +57,8 @@ func (cmd *readHeaderCommand) parseResult(ifc command, conn *Connection) error {
 	resultCode := cmd.dataBuffer[13] & 0xFF
 
 	if resultCode == 0 {
-		generation := int(Buffer.BytesToInt32(cmd.dataBuffer, 14))
-		expiration := TTL(int(Buffer.BytesToInt32(cmd.dataBuffer, 18)))
+		generation := int(uint32(Buffer.BytesToInt32(cmd.dataBuffer, 14)))
+		expiration := TTL(int(uint32(Buffer.BytesToInt32(cmd.dataBuffer, 18))))
 		cmd.record = newRecord(cmd.node, cmd.key, nil, nil, generation, expiration)
 	} else {
 		if ResultCode(resultCode) == KEY_NOT_FOUND_ERROR {
